FWF-4352: [Bugfix] Sonar fix
