+++ .github/workflows/forms-flow-data-analysis-api-ci.yml
@@ -69,6 +69,7 @@ jobs:
       MODEL_ID: ${{ secrets.MODEL_ID }}
       SKIP_IN_CI: "True"
       USE_DOCKER_MOCK: "True"
+      JWT_OIDC_TEST_PRIVATE_KEY_PEM: ${{ secrets.JWT_OIDC_TEST_PRIVATE_KEY_PEM }}
 
     runs-on: ubuntu-20.04
