+++ forms-flow-api-utils/src/formsflow_api_utils/utils/__init__.py
@@ -59,3 +59,4 @@ from .util import (
 from .caching import Cache
 from .sentry import init_sentry
 from .formio import generate_formio_patch_request
+from .test_utils import get_token
